Added Users and provided OAuth github functionality
The page is at the stage of being able to redirect the user to github for login upon entering /auth/github. Callback is also implemented with ~ok decoupled implementation. This is the part of the course where I know enough about the stuff to try to make improvements. Besides, the material is for 1.2 so since I'm using 1.3 there is a lot to change C: